Added the Lesson Test70
