+++ src/main/java/kg/peaksoft/peaksoftlmsab4/PeaksoftlmsAB4Application.java
@@ -31,7 +31,7 @@ public class PeaksoftlmsAB4Application {
         System.out.println("Welcome colleagues, project name is Peaksoftlms-A!");
     }
 
-//    @PostConstruct
+    @PostConstruct
     public void init() {
         UserEntity user = new UserEntity();
         user.setFirstName("Admin");
