+++ src/main/java/kg/peaksoft/peaksoftlmsab4/exception/AlreadyExistsException.java
@@ -1,8 +1,5 @@
 package kg.peaksoft.peaksoftlmsab4.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
 public class AlreadyExistsException extends RuntimeException {
 
     public AlreadyExistsException() {
