+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/mapper/OptionMapper.java
@@ -1,9 +1,9 @@
 package kg.peaksoft.peaksoftlmsab4.model.mapper;
 
+import kg.peaksoft.peaksoftlmsab4.api.payload.OptionRequest;
 import kg.peaksoft.peaksoftlmsab4.api.payload.OptionResponse;
-
 import kg.peaksoft.peaksoftlmsab4.model.entity.OptionEntity;
-
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
@@ -12,8 +12,28 @@ import java.util.List;
 
 @Component
 @Slf4j
-public class OptionViewMapper {
+@RequiredArgsConstructor
+public class OptionMapper {
 
+
+
+    public OptionEntity create(OptionRequest optionRequest) {
+
+        OptionEntity option = new OptionEntity();
+
+        option.setAnswer(optionRequest.getAnswer());
+        option.setIsTrue(optionRequest.getIsTrue());
+        return option;
+
+    }
+
+
+
+    public OptionEntity update(OptionEntity option,OptionRequest optionRequest){
+        option.setIsTrue(optionRequest.getIsTrue());
+        option.setAnswer(optionRequest.getAnswer());
+        return option;
+      }
 
     public OptionResponse viewOption(OptionEntity option) {
         if (option == null) {
