+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/mapper/QuestionMapper.java
@@ -2,9 +2,9 @@ package kg.peaksoft.peaksoftlmsab4.model.mapper;
 
 import kg.peaksoft.peaksoftlmsab4.api.payload.OptionRequest;
 import kg.peaksoft.peaksoftlmsab4.api.payload.QuestionRequest;
+import kg.peaksoft.peaksoftlmsab4.api.payload.QuestionResponse;
 import kg.peaksoft.peaksoftlmsab4.model.entity.OptionEntity;
 import kg.peaksoft.peaksoftlmsab4.model.entity.QuestionEntity;
-import kg.peaksoft.peaksoftlmsab4.service.OptionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -15,15 +15,16 @@ import java.util.List;
 @Component
 @Slf4j
 @RequiredArgsConstructor
-public class QuestionEditMapper {
-    private final OptionEditMapper optionEditMapper;
+public class QuestionMapper {
+    private final OptionMapper optionMapper;
+
     public QuestionEntity create(QuestionRequest questionRequest) {
          List<OptionEntity>optionEntities=new ArrayList<>();
         QuestionEntity question = new QuestionEntity();
         question.setQuestion(questionRequest.getQuestion());
         question.setQuestionType(questionRequest.getQuestionType());
         for (OptionRequest o: questionRequest.getOptions()) {
-            optionEntities.add(optionEditMapper.create(o));
+            optionEntities.add(optionMapper.create(o));
         }
         question.setOptions(optionEntities);
 
@@ -34,5 +35,30 @@ public class QuestionEditMapper {
         question.setQuestion(questionRequest.getQuestion());
         return question;
 
+    }
+
+    public QuestionResponse viewQuestion(QuestionEntity question) {
+
+        if (question == null) {
+            log.error("The question db is null!");
+            return null;
+        }
+        QuestionResponse questionResponse = new QuestionResponse();
+        if (question.getId() != null) {
+            questionResponse.setId(question.getId());
+        }
+        questionResponse.setQuestion(question.getQuestion());
+        questionResponse.setQuestionType(question.getQuestionType());
+        questionResponse.setOptions(optionMapper.viewOptions(question.getOptions()));
+        return questionResponse;
+    }
+
+    public List<QuestionResponse> viewQuestions(List<QuestionEntity> questions) {
+
+        List<QuestionResponse> questionResponses = new ArrayList<>();
+        for (QuestionEntity q : questions) {
+            questionResponses.add(viewQuestion(q));
+        }
+        return questionResponses;
     }
 }
