+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/mapper/TestMapper.java
@@ -3,17 +3,13 @@ package kg.peaksoft.peaksoftlmsab4.model.mapper;
 import kg.peaksoft.peaksoftlmsab4.api.payload.OptionRequest;
 import kg.peaksoft.peaksoftlmsab4.api.payload.QuestionRequest;
 import kg.peaksoft.peaksoftlmsab4.api.payload.TestRequest;
+import kg.peaksoft.peaksoftlmsab4.api.payload.TestResponse;
 import kg.peaksoft.peaksoftlmsab4.exception.BadRequestException;
-import kg.peaksoft.peaksoftlmsab4.exception.NotFoundException;
-import kg.peaksoft.peaksoftlmsab4.model.entity.OptionEntity;
 import kg.peaksoft.peaksoftlmsab4.model.entity.QuestionEntity;
 import kg.peaksoft.peaksoftlmsab4.model.entity.TestEntity;
 import kg.peaksoft.peaksoftlmsab4.model.enums.QuestionType;
-import kg.peaksoft.peaksoftlmsab4.repository.OptionRepository;
-import kg.peaksoft.peaksoftlmsab4.repository.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -22,38 +18,11 @@ import java.util.List;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class TestEditMapper {
-    private final QuestionEditMapper editMapper;
+public class TestMapper {
+    private final QuestionMapper mapper;
+
 
     public TestEntity create(TestRequest testRequest) {
-//        List<QuestionEntity> questions = new ArrayList<>();
-//
-//        TestEntity test = new TestEntity();
-//        test.setTestName(testRequest.getTestName());
-//        int counter = 0;
-//        for (QuestionRequest q : testRequest.getQuestions()) {
-//
-//            if (q.getQuestionType() == QuestionType.ONE) {
-//                for (OptionRequest o : q.getOptions()) {
-//                    if (o.getIsTrue()) {
-//                        counter++;
-//                    }
-//                }
-//                if (counter > 1) {
-//                    throw new BadRequestException("You can't choose multiple variants");
-//                } else {
-//                    questions.add(editMapper.create(q));
-//                    test.setQuestions(questions);
-//                    return test;
-//                }
-//
-//            }else
-//                questions.add(editMapper.create(q));
-//                test.setQuestions(questions);
-//                return test;
-//        }
-//        return null;
-//    }
         List<QuestionEntity> questions = new ArrayList<>();
         TestEntity test = new TestEntity();
         test.setTestName(testRequest.getTestName());
@@ -68,18 +37,41 @@ public class TestEditMapper {
                 if (counter > 1) {
                     throw new BadRequestException("You can't choose multiply answer");
                 } else {
-                    questions.add(editMapper.create(q));
+                    questions.add(mapper.create(q));
                     test.setQuestions(questions);
                     return test;
                 }
 
             } else
-                questions.add(editMapper.create(q));
+                questions.add(mapper.create(q));
             test.setQuestions(questions);
             return test;
 
         }
         return null;
+    }
+
+    public TestResponse viewTest(TestEntity test) {
+        if (test == null) {
+            log.error("The test db is null!");
+            return null;
+        }
+        TestResponse testResponse = new TestResponse();
+        if (test.getId() != null) {
+            testResponse.setId(test.getId());
+        }
+        testResponse.setTestName(test.getTestName());
+        testResponse.setQuestions(mapper.viewQuestions(test.getQuestions()));
+        return testResponse;
+
+    }
+
+    public List<TestResponse> viewTests(List<TestEntity> tests) {
+        List<TestResponse> testResponses = new ArrayList<>();
+        for (TestEntity t : tests) {
+            testResponses.add(viewTest(t));
+        }
+        return testResponses;
     }
