+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/mapper/TestStudentMapper.java
@@ -1,6 +1,7 @@
 package kg.peaksoft.peaksoftlmsab4.model.mapper;
 
 import kg.peaksoft.peaksoftlmsab4.api.payload.TestStudentRequest;
+import kg.peaksoft.peaksoftlmsab4.api.payload.TestStudentResponse;
 import kg.peaksoft.peaksoftlmsab4.exception.NotFoundException;
 import kg.peaksoft.peaksoftlmsab4.model.entity.OptionEntity;
 import kg.peaksoft.peaksoftlmsab4.model.entity.TestStudentEntity;
@@ -11,16 +12,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
 
-public class TestStudentEditMapper {
+public class TestStudentMapper {
     private final OptionRepository optionRepository;
-    private final TestStudentRepository studentRepository;
+
 
 
     public List<TestStudentEntity> createResult(TestStudentRequest testStudentRequest) {
@@ -28,24 +28,16 @@ public class TestStudentEditMapper {
 
         List<OptionEntity> options = optionRepository.findAll();
 
-
-        System.out.println(options);
-
         List<TestStudentEntity> resultEntities = new ArrayList<>();
 
         TestStudentEntity testStudent = new TestStudentEntity();
 
-        Iterator<TestStudentEntity>iterator= resultEntities.iterator();
         for (OptionEntity o : options) {
-
-
             testStudent.setAnswer(o.getAnswer());
             testStudent.setIsTrue(o.getIsTrue());
-resultEntities.add(testStudent);
-
-
+            resultEntities.add(testStudent);
         }
-return resultEntities;
+        return resultEntities;
     }
 
 
@@ -62,5 +54,23 @@ return resultEntities;
                             String.format("Question with id = %s does not exists", optionId)
                     );
                 });
+    }
+
+    public TestStudentResponse viewTestResult(TestStudentEntity resultEntity) {
+        TestStudentResponse testResultResponse = new TestStudentResponse();
+        testResultResponse.setId(resultEntity.getId());
+        testResultResponse.setAnswer(resultEntity.getAnswer());
+
+
+        return testResultResponse;
+    }
+
+    public List<TestStudentResponse> viewTestResults(List<TestStudentEntity> testStudentEntities) {
+
+        List<TestStudentResponse> testResultResponses = new ArrayList<>();
+        for (TestStudentEntity o : testStudentEntities) {
+            testResultResponses.add(viewTestResult(o));
+        }
+        return testResultResponses;
     }
 }
