+++ src/main/java/kg/peaksoft/peaksoftlmsab4/service/serviceImpl/TestServiceImpl.java
@@ -4,8 +4,7 @@ import kg.peaksoft.peaksoftlmsab4.api.payload.TestRequest;
 import kg.peaksoft.peaksoftlmsab4.api.payload.TestResponse;
 import kg.peaksoft.peaksoftlmsab4.exception.NotFoundException;
 import kg.peaksoft.peaksoftlmsab4.model.entity.TestEntity;
-import kg.peaksoft.peaksoftlmsab4.model.mapper.TestEditMapper;
-import kg.peaksoft.peaksoftlmsab4.model.mapper.TestViewMapper;
+import kg.peaksoft.peaksoftlmsab4.model.mapper.TestMapper;
 import kg.peaksoft.peaksoftlmsab4.repository.TestRepository;
 import kg.peaksoft.peaksoftlmsab4.service.TestService;
 import lombok.RequiredArgsConstructor;
@@ -17,27 +16,29 @@ import java.util.List;
 @RequiredArgsConstructor
 public class TestServiceImpl implements TestService {
 
-    private final TestViewMapper viewMapper;
+
     private final TestRepository repository;
-    private final TestEditMapper editMapper;
+    private final TestMapper testMapper;
 
     @Override
     public TestResponse create(TestRequest testRequest) {
-        return viewMapper.viewTest(repository.save(editMapper.create(testRequest)));
+        return testMapper.viewTest(repository.save(testMapper.create(testRequest)));
     }
 
     @Override
     public TestResponse update(Long id, TestRequest testRequest) {
         TestEntity test = repository.findById(id).orElseThrow(() -> new NotFoundException(
                 String.format("test not found this=%s", id)));
-        editMapper.update(test,testRequest);
+        testMapper.update(test,testRequest);
 
-        return viewMapper.viewTest(repository.save(test));
+        return testMapper.viewTest(repository.save(test));
     }
 
     @Override
     public TestResponse findById(Long id) {
-        return viewMapper.viewTest(repository.findById(id).get());
+        return testMapper.viewTest(repository.findById(id).orElseThrow(()->new NotFoundException(
+                String.format("test not found this=%s",id)
+        )));
     }
 
     @Override
@@ -45,11 +46,11 @@ public class TestServiceImpl implements TestService {
         TestEntity test = repository.findById(id).orElseThrow(() -> new NotFoundException(
                 String.format("test not found this=%s", id)));
         repository.deleteById(id);
-        return viewMapper.viewTest(test);
+        return testMapper.viewTest(test);
     }
 
     @Override
     public List<TestResponse> findAll() {
-        return viewMapper.viewTests(repository.findAll());
+        return testMapper.viewTests(repository.findAll());
     }
 }
